Removed old code, relocated unused function.
Code pertaining to the removal of field groups from the parent theme was a left over reminent of when this theme was connected to UDS-WordPress. The move to Pitchfork made these obsolete.

Code pertaining to data validation in ACF was moved to a separate file and not included in the stack. This preserves the idea for later when the feature in ACF is finally fixed?